make attributes optional

+++ Aramex/Models/MechnicalSite.cs
@@ -11,11 +11,11 @@ namespace Aramex.Models
         public string ComponentName { get; set; }
         public int FCUAddress { get; set; }
         public string ServedArea { get; set; }
-        public double? RunHoursWork { get; set; }
-        public double? PreventiveMaintainanceRun { get; set; }
-        public double? PreventiveMaintainanceOverdue { get; set; }
+        public int? RunHoursWork { get; set; }
+        public int? PreventiveMaintainanceRun { get; set; }
+        public int? PreventiveMaintainanceOverdue { get; set; }
         public string RunName { get; set; }
         public string OverdueName { get; set; }
-        public double? TripHours { get; set; }
+        public int? TripHours { get; set; }
     }
 }
